accounts/requirements.txt
attrs==23.1.0
click==7.1.2
dnspython==2.3.0
Flask==1.1.4
Flask-Cors==3.0.10
grpcio==1.54.2
grpcio-tools==1.54.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
more-itertools==9.1.0
packaging==23.1
pluggy==0.13.1
protobuf==4.23.2
py==1.11.0
pymongo==4.3.3
pytest==5.4.3
six==1.16.0
wcwidth==0.2.6
Werkzeug==1.0.1
requests
dotmap
python-dotenv==1.0.0